constants.py added for global constants variables
